Split filesystem from s3 into s3fs
