+++ distributed/tests/test_s3fs.py
@@ -1,0 +1,137 @@
+# -*- coding: utf-8 -*-
+import pytest
+from distributed.s3fs import S3FileSystem
+from distributed.s3 import seek_delimiter
+from distributed.utils_test import slow
+import moto
+
+# These get mirrored on s3://distributed-test/
+test_bucket_name = 'distributed-test'
+files = {'test/accounts.1.json':  (b'{"amount": 100, "name": "Alice"}\n'
+                                   b'{"amount": 200, "name": "Bob"}\n'
+                                   b'{"amount": 300, "name": "Charlie"}\n'
+                                   b'{"amount": 400, "name": "Dennis"}\n'),
+         'test/accounts.2.json':  (b'{"amount": 500, "name": "Alice"}\n'
+                                   b'{"amount": 600, "name": "Bob"}\n'
+                                   b'{"amount": 700, "name": "Charlie"}\n'
+                                   b'{"amount": 800, "name": "Dennis"}\n')}
+
+csv_files = {'2014-01-01.csv': (b'name,amount,id\n'
+                                b'Alice,100,1\n'
+                                b'Bob,200,2\n'
+                                b'Charlie,300,3\n'),
+             '2014-01-02.csv': (b'name,amount,id\n'),
+             '2014-01-03.csv': (b'name,amount,id\n'
+                                b'Dennis,400,4\n'
+                                b'Edith,500,5\n'
+                                b'Frank,600,6\n')}
+
+@pytest.yield_fixture
+def s3():
+    # could do with a bucket with write privileges.
+    yield S3FileSystem(anon=True)
+
+
+def test_s3_file_access(s3):
+    fn = 'distributed-test/nested/file1'
+    data = b'hello\n'
+    assert s3.cat(fn) == data
+    assert s3.head(fn, 3) == data[:3]
+    assert s3.tail(fn, 3) == data[-3:]
+    assert s3.tail(fn, 10000) == data
+
+
+def test_s3_file_info(s3):
+    fn = 'distributed-test/nested/file1'
+    data = b'hello\n'
+    assert fn in s3.walk('distributed-test')
+    assert s3.exists(fn)
+    assert not s3.exists(fn+'another')
+    assert s3.info(fn)['Size'] == len(data)
+    with pytest.raises((OSError, IOError)):
+        s3.info(fn+'another')
+
+def test_du(s3):
+    d = s3.du(test_bucket_name, deep=True)
+    assert all(isinstance(v, int) and v >= 0 for v in d.values())
+    assert 'distributed-test/nested/file1' in d
+
+    assert s3.du(test_bucket_name + '/test/', total=True) ==\
+           sum(map(len, files.values()))
+
+
+def test_s3_ls(s3):
+    fn = 'distributed-test/nested/file1'
+    assert fn not in s3.ls('distributed-test/')
+    assert fn in s3.ls('distributed-test/nested/')
+    assert fn in s3.ls('distributed-test/nested')
+    assert s3.ls('s3://distributed-test/nested/') == s3.ls('distributed-test/nested')
+
+
+def test_s3_ls_detail(s3):
+    L = s3.ls('distributed-test/nested', detail=True)
+    assert all(isinstance(item, dict) for item in L)
+
+
+def test_s3_glob(s3):
+    fn = 'distributed-test/nested/file1'
+    assert fn not in s3.glob('distributed-test/')
+    assert fn not in s3.glob('distributed-test/*')
+    assert fn in s3.glob('distributed-test/nested')
+    assert fn in s3.glob('distributed-test/nested/*')
+    assert fn in s3.glob('distributed-test/nested/file*')
+    assert fn in s3.glob('distributed-test/*/*')
+
+
+def test_get_list_of_summary_objects(s3):
+    L = s3.ls(test_bucket_name + '/test')
+
+    assert len(L) == 2
+    assert [l.lstrip(test_bucket_name).lstrip('/') for l in sorted(L)] == sorted(list(files))
+
+    L2 = s3.ls('s3://' + test_bucket_name + '/test')
+
+    assert L == L2
+
+
+def test_read_keys_from_bucket(s3):
+    for k, data in files.items():
+        file_contents = s3.cat('/'.join([test_bucket_name, k]))
+        assert file_contents == data
+
+    assert (s3.cat('/'.join([test_bucket_name, k])) ==
+            s3.cat('s3://' + '/'.join([test_bucket_name, k])))
+
+
+@slow
+def test_seek_delimiter(s3):
+    fn = 'test/accounts.1.json'
+    data = files[fn]
+    with s3.open('/'.join([test_bucket_name, fn])) as f:
+        seek_delimiter(f, b'}', 0)
+        assert f.tell() == 0
+        f.seek(1)
+        seek_delimiter(f, b'}', 5)
+        assert f.tell() == data.index(b'}') + 1
+        seek_delimiter(f, b'\n', 5)
+        assert f.tell() == data.index(b'\n') + 1
+        f.seek(1, 1)
+        ind = data.index(b'\n') + data[data.index(b'\n')+1:].index(b'\n') + 1
+        seek_delimiter(f, b'\n', 5)
+        assert f.tell() == ind + 1
+
+
+def test_read_s3_block(s3):
+    import io
+    data = files['test/accounts.1.json']
+    lines = io.BytesIO(data).readlines()
+    path = 'distributed-test/test/accounts.1.json'
+    assert s3.read_block(path, 1, 35, b'\n') == lines[1]
+    assert s3.read_block(path, 0, 30, b'\n') == lines[0]
+    assert s3.read_block(path, 0, 35, b'\n') == lines[0] + lines[1]
+    assert s3.read_block(path, 0, 5000, b'\n') == data
+    assert len(s3.read_block(path, 0, 5)) == 5
+    assert len(s3.read_block(path, 4, 5000)) == len(data) - 4
+    assert s3.read_block(path, 5000, 5010) == b''
+
+    assert s3.read_block(path, 5, None) == s3.read_block(path, 5, 1000)
